requirements.txt
discord.py
pynacl